% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genSWATdates.R
\name{genSWATdates}
\alias{genSWATdates}
\title{generates a vector of dates in the swat model IO format}
\usage{
genSWATdates(startDate, endDate)
}
\arguments{
\item{startDate}{- Start date for output vector, supports common date formats}

\item{endDate}{- End date for output vector, supports common date formats}
}
\value{
returns a generated time series of swat IO format dates
}
\description{
generates a vector of dates in the swat model IO format
}
\examples{
startDate = '2000-12-28'
endDate = '2000-12-29'
genSWATdates(startDate=startDate, endDate=endDate)
}

