% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSWATdates.R
\name{getSWATdates}
\alias{getSWATdates}
\title{Converts a date time series to SWAT IO format dates}
\usage{
getSWATdates(dates, dateformat = "yyyy-mm-dd")
}
\arguments{
\item{dates}{- Input vector of dates in as Date or character class.}

\item{dateformat}{- Format of input dates. Can be 'yyyy-mm-dd', 'mm-dd-yyyy', 'mm/dd/yyyy', 'yyyy/mm/dd','dd/mm/yyy', or 'dd-mm-yyy'. Default is 'yyyy-mm-dd'.}
}
\value{
returns a time series of swat IO format dates from the input
}
\description{
Converts a date time series to SWAT IO format dates
}
\examples{
dates = c('2000-12-28', '2000-12-29')
getSWATdates(dates)
}

